backend/requirements.txt
prisma
gunicorn
fastapi
uvicorn
python-dotenv
python-jose
passlib
python-multipart
bcrypt == 4.0.1
requests
boto3
botocore